refactor: prune duplicated MANIFEST and requirements
