+++ MANIFEST.client-cli.in
@@ -1,11 +1,6 @@
-include CHANGELOG.md
-include LICENSE
-include README.rst
-graft src/ai/backend/client/cli
-graft src/ai/backend/client/output
-include src/ai/backend/client/__init__.py
-include src/ai/backend/client/compat.py
-include src/ai/backend/client/config.py
-include src/ai/backend/client/exceptions.py
-include src/ai/backend/client/session.py
-include src/ai/backend/client/types.py
+exclude src/ai/backend/client/auth.py
+exclude src/ai/backend/client/helper.py
+exclude src/ai/backend/client/load_balancing.py
+exclude src/ai/backend/client/pagination.py
+exclude src/ai/backend/client/request.py
+exclude src/ai/backend/client/utils.py
