+++ MANIFEST.client.in
@@ -1,19 +1,2 @@
-include CHANGELOG.md
-include LICENSE
-include README.rst
-graft src/ai/backend/client/func
-include src/ai/backend/client/__init__.py
-include src/ai/backend/client/auth.py
-include src/ai/backend/client/compat.py
-include src/ai/backend/client/config.py
-include src/ai/backend/client/exceptions.py
-include src/ai/backend/client/helper.py
-include src/ai/backend/client/load_balancing.py
-include src/ai/backend/client/pagination.py
-include src/ai/backend/client/py.typed
-include src/ai/backend/client/request.py
-include src/ai/backend/client/session.py
-include src/ai/backend/client/test_utils.py
-include src/ai/backend/client/types.py
-include src/ai/backend/client/utils.py
-include src/ai/backend/client/versioning.py
+prune src/ai/backend/client/cli
+prune src/ai/backend/client/output
