+++ MANIFEST.in
@@ -1,4 +1,6 @@
 include CHANGELOG.md
 include LICENSE
 include README.rst
-graft src/ai/backend/client
+graft docs
+graft src/ai/backend
+include setup_orig.py
